jetls-client: Split the client into per-component modules and unit-test the transport layer (#858)
* jetls-client: Split the client into per-component modules

`jetls-client.ts` had grown to roughly 900 lines mixing the extension
entry points with the status bar, configuration handling, the server
lifecycle, and update checks, while only small vscode-free helpers
lived in separate files. This made the layout look arbitrary: one tiny
utility (`CoalescingTaskRunner`) had its own file while everything
else shared one.

This commit splits the client per component: `jetls-client.ts` keeps
only `activate`/`deactivate` (the esbuild entry name is preserved),
`server-lifecycle.ts` owns the server start/restart/shutdown flow,
`status-bar.ts` wraps the startup status bar in a `StartupStatusBar`
class, `server-config.ts` reads and compares the server configuration,
and `update-check.ts` hosts the version update checks. It also renames
`server-startup.ts` to `preflight.ts` and `CoalescingTaskRunner.ts` to
`coalescing-task-runner.ts`, following the file naming convention.

The moves are verbatim except for the status bar: the `deactivating`
guard inside the old `showServerStartupStatus` is now a `suppress()`
method on `StartupStatusBar`, called at the start of `deactivate`.
`CoalescingTaskRunner` stays in its own vscode-free module because
modules imported from `test/*.test.ts` are bundled without `vscode`
as an external, so unit-tested code cannot depend on it.


* jetls-client: Extract socket/pipe transport into a testable module

The socket and pipe connection establishment lived inline in
`server-lifecycle.ts` as hand-written async state machines: the
`<JETLS-PORT>` handshake, the pipe connect-back dance with its
exit-before-connect and cleanup invariants, and the startup timeout
that re-arms when precompilation output is detected. None of it had
test coverage, because `server-lifecycle.ts` depends on `vscode` and
modules imported from `test/*.test.ts` are bundled without `vscode`
as an external.

This commit moves that logic verbatim into a new vscode-free
`transport.ts` exposing `connectSocketTransport` and
`connectPipeTransport`. Following the `VersionPreflight` pattern,
`TransportOptions` injects `spawnProcess`, `appendLine`,
`onPrecompiling`, and `onProcessError`; the spawn-error popup UI
stays in `server-lifecycle.ts` behind the `onProcessError` callback.
The returned `TransportStreams` type is structurally compatible with
vscode-languageclient's `StreamInfo`, so the module does not need to
import it. Log messages and behavior are unchanged.

The shared `FakeChildProcess` helper moves from `preflight.test.ts`
to `test/fake-child-process.ts` and now emits `exit` before `close`,
since the transport watches `exit`. The new `transport.test.ts` adds
ten tests that fake only the child process and use real loopback TCP
and Unix domain sockets, covering the port handshake, exit without a
port, startup timeouts and the precompilation re-arm, connection
refusal, exit before the pipe connects, socket file cleanup, and
spawn error propagation.


---------
